requirements.txt
fastapi
uvicorn[standard]
scikit-learn
joblib
pandas
numpy
python-jose[cryptography]
passlib
bcrypt==3.2.0
python-multipart
email-validator
SQLAlchemy
psycopg2-binary